+++ game/systems/s_light.cpp
@@ -8,16 +8,13 @@
 
 #include "ftime.h"
 #include "r_opengl.h"
+#include "r_renderer.h"
 
 #define LIGHT_PERIOD 5
 #define LIGHT_VARIATION 5
 #define LIGHT_CENTER_RADIUS 0.25
 
 using namespace Verse;
-
-
-
-#include "log.h"
 
 namespace {
     std::map<EntityID, glm::vec4> light_sources;
@@ -35,13 +32,14 @@ void System::Light::render(Config &c, ui8 pid) {
         if (tex != nullptr)
             light_sources[e] += glm::vec4(tex->transform.x, tex->transform.y, 0, 0);
         
-        light_sources[e].x /= c.resolution.x;
-        light_sources[e].y = 1.0f - light_sources[e].y / c.resolution.y;
-        light_sources[e].z /= c.resolution.x;
-        light_sources[e].w /= c.resolution.x;
+        light_sources[e].x /= c.resolution.x + 2.0f*BORDER_WIDTH;
+        light_sources[e].y = 1.0f - light_sources[e].y / (c.resolution.y + 2.0f*BORDER_WIDTH);
+        light_sources[e].z /= c.resolution.x + 2.0f*BORDER_WIDTH;
+        light_sources[e].w /= c.resolution.x + 2.0f*BORDER_WIDTH;
         
-        light_sources[e].x += (0.5 - ((float)floor(c.active_camera->pos.x) / c.resolution.x));
-        light_sources[e].y -= (0.5 - ((float)floor(c.active_camera->pos.y) / c.resolution.y));
+        light_sources[e].x += (0.5f * c.resolution.x - BORDER_WIDTH - (float)floor(c.active_camera->pos.x)) / (c.resolution.x + 2.0f*BORDER_WIDTH);
+        light_sources[e].y -= (0.5f * c.resolution.y - BORDER_WIDTH - (float)floor(c.active_camera->pos.y)) / (c.resolution.y + 2.0f*BORDER_WIDTH);
+        
     }
     
     std::vector<glm::vec4> light_data;
@@ -51,7 +49,7 @@ void System::Light::render(Config &c, ui8 pid) {
     
     glUniform4fv(glGetUniformLocation(pid, "light"), (int)(light_sources.size()), reinterpret_cast<GLfloat *>(light_data.data()));
     glUniform1i(glGetUniformLocation(pid, "light_size"), (int)(light_sources.size()));
-    glUniform1f(glGetUniformLocation(pid, "light_distortion"), (float)c.resolution.x / (float)c.resolution.y);
+    glUniform1f(glGetUniformLocation(pid, "light_distortion"), (float)(c.resolution.x + 2*BORDER_WIDTH) / (float)(c.resolution.y + 2*BORDER_WIDTH));
 }
 
 void System::Light::clean() {